Fix for the Scene Window Zoom Bug
This removes a duplicate code block that was causing the rendering rect
to shrink twice as fast as the window when zooming in which caused
sprites to disappear.  This fixes issue #53 and was originally suggested
by Simon Love.  I compared the file with the original in TGB to make
sure we deleted the correct code block.  This has been tested by Simon
and myself in our projects and in the toys.