Updated MoveTo
This causes MoveTo to function based on position rather than time.  This
corrects problems with objects teleporting when they don't reach their
position before they are supposed to.  This was most commonly happening
because of the speed cap imposed by box2D.  The new method works the
same as before with the re-introduction of the margin as the optional
last parameter.  The new function is also more taxing on the engine
although it is unlikely that any loss in frame rate will be seen unless
hundreds of objects are being moved at the same time.