CompositeSprite Animation Bug
Previously, CompositeSprites would continue to animate even when the
scene was paused.  This is because the ImageFrameProvider for each
sprite would receive the tick event and process it independently of the
Scene.  I turned this off.  Now when the Scene processes a
CompositeSprite, each sprite in the composite is given the chance to
advance its animation.  If the Scene is paused, then there is no
animation.